Moved ReleaseWithSources to a top-level class, moved comparator.
Issue: eclipse#426.

Being a top-level class simplifies usage of ReleaseWithSources.

The comparator for this class is now contained in SW360Exporter
because this class is in control of the sort order.
